+++ CHANGELOG.rst
@@ -13,6 +13,9 @@ New Features
   :py:func:`~climpred.predictability_horizon.predictability_horizon` based on condition.
   (:issue:`46`, :pr:`521`) `Aaron Spring`_.
 
+Bug fixes
+---------
+- :py:meth:`~climpred.classes.PredictionEnsemble.smooth` now carries lead.attrs (:issue: `527`, pr:`521`) `Aaron Spring`_.
 
 climpred v2.1.1 (2020-10-13)
 ============================